Split Wearables strip action owner
